chore: rename to config
